requirements.txt
pytest>=7.0.0
requests>=2.28.0
numpy>=1.24.0
sympy>=1.12.0
networkx>=3.0
black>=24.10.0
ruff>=0.8.1
python-dotenv>=1.0
beautifulsoup4>=4.12.2
markdownify>=0.11
pytz>=2024.2
